Tests/LibWeb/Text/expected/IntersectionObserver/scroll-margin-expands-scrollport.txt

no-margin: ratio=0.1875
with-margin: ratio=0.1875
